Moved container build logic into sbt plugin
