+++ project/BlendedDemoMgmt.scala
@@ -1,9 +1,8 @@
+import blended.sbt.container.BlendedContainerPlugin.autoImport._
+import blended.sbt.feature.BlendedFeaturePlugin.autoImport._
 import com.typesafe.sbt.packager.universal.UniversalPlugin.autoImport._
 import sbt.Keys._
 import sbt._
-import BlendedContainer._
-
-import blended.sbt.feature.BlendedFeaturePlugin.autoImport._
 
 object BlendedDemoMgmt extends ProjectFactory {
 
@@ -14,6 +13,7 @@ object BlendedDemoMgmt extends ProjectFactory {
   ) {
 
     override def settings: Seq[sbt.Setting[_]] = super.settings ++ Seq(
+      blendedVersion := Blended.blendedVersion,
       materializeDebug := false,
       materializeExtraDeps := {
         val file = (BlendedDemoMgmtResources.project / Universal / packageBin).value
